Extended color support, small fixes
- Errors now use the sublimeline.illegal style, it's more explicit and
consistent with naming of other styles (and allowed refactoring of code)
- Added "sublimelint_fill_outlines" setting, if true the offending lines
are drawn filled with the background color instead of outlined. Great if
you want to make it glaringly obvious that you have errors. - Updated
the documentation on the colors accordingly. - The invalid.* color
should actually be the background, not the foreground. Updated the
sample styles accordingly. - Fixed pep8 errors in sublime_pylint.py and
jshint.py. SublimeLint should pass its own tests!