replace "show_outlines" with "outline_style"
The new setting can be either "outline" (default, draw outline
boxes), "fill" (fill the boxes), or "none" (nothing is shown). For
compatibility reasons, if "outline" is used and the deprecated
setting "fill_outlines" is set, "outline_style" is switched to
"fill".

This is so that users using "fill_outlines" still have the same
behaviour even if "outline_style" is set to "outline" in the
default settings, so the behaviour doesn't change on upgrade. The
logic is somewhat convoluted and should be removed once we don't
support the deprecated "fill_outlines" any longer.