Explicitly importing modules from libs
sys.path appears to ignore individual paths with unicode characters.
This means that this lib_path will be ignored for Windows 7 users
with non-ascii characters in their username (thus as their home
directory).

As a fix for the Windows 7 lib path issue (#181), the individual
modules in the `libs` folder can be explicitly imported. This
obviously doesn't scale well, but may be a necessary evil until ST2
upgrades its internal Python.